Adjust queue and semaphore unit tests for full coverage of xQueueGenericGetStaticBuffers
Covers xQueueGenericGetStaticBuffers, xSemaphoreGetStaticBuffer, and xQueueGetStaticBuffers
