+++ FreeRTOS/Test/CMock/queue/static/queue_get_static_buffers_utest.c
@@ -1,0 +1,1 @@
+../generic/queue_get_static_buffers_utest.c
